getProfileTraits moved to base class
